cleanup, fixing broken commands from modularization effort, etc.

+++ commands/cleanSingle.js
@@ -6,7 +6,14 @@ module.exports={
         this.db.get('SELECT * FROM applicants WHERE tag = $tag', {
             $tag: args.tag
         }, (err, row) => {
-            if (err) { console.log('oops', err); }
+            if (err) { 
+                console.log('oops', err); 
+                return callback();
+            }
+            if(!row){
+                console.log('there isn\'t a test with that tag');
+                return callback();
+            }
             request({
                 method: 'DELETE',
                 url: row.apiurl || 'wtf',
@@ -14,20 +21,19 @@ module.exports={
                     'User-Agent': 'winnow-code-test',
                     'Authorization': `token ${this.config.privToken}`
                 }
-            },
-          (err) => {
-              if (err) {
-                  console.log('ooops', err);
-                  rimraf.sync('tmp');
-              }
+            },err => {
+                if (err) {
+                    console.log('ooops', err);
+                    rimraf.sync('tmp');
+                }
               // console.log('response: ', res.status, res.body);
-              this.db.run('DELETE FROM applicants WHERE tag = $tag', {$tag: args.tag}, function(err){
-                  if(err){
-                      console.log(err);
-                  }
-                  return callback();
-              });
-          });
+                this.db.run('DELETE FROM applicants WHERE tag = $tag', {$tag: args.tag}, function(err){
+                    if(err){
+                        console.log(err);
+                    }
+                    return callback();
+                });
+            });
         });    
     }
 };
